+++ tests/test_architectures.py
@@ -1,19 +1,10 @@
-from pathlib import Path
-
 import numpy as np
-import pytest
 
 from lr_face.models import Architecture
 from lr_face.utils import fix_tensorflow_rtx
+from tests.conftest import skip_on_github
 
 fix_tensorflow_rtx()
-
-
-def skip_on_github(func):
-    return pytest.mark.skipif(
-        str(Path.home()) == '/home/runner',
-        reason="Fails on Github because model weights don't exist"
-    )(func)
 
 
 @skip_on_github
